Refactored agent.py to abstract_agent.py
