+++ test/integration_tests/question.py
@@ -1,6 +1,6 @@
 from parameterized import parameterized
 from .integration_base import BaseIntegrationTest
-from agent.agents.agent import apology_string
+from agent.agents.abstract_agent import apology_string
 
 class TestQuestion(BaseIntegrationTest):
     launch_sequence = ['let\'s chat']
